Make nvmish less noisy
Jonah's trying to run the developer start and is currently blocked by the `return 1` that happens by default when invoking `nvmish` bash function.

This PR stops returning non-zero exit code by default when cd'ing into a directory. It also creates a debug function (but only uses it one place).

This _might_ be a major version bump since it changes the default behavior.