Specify "SubjectAltName" in self-signed SSL cert.
Chrome 58 (among others) requires that self-signed certificates specify the hostname(s)
to which they apply in the `SubjectAltName` field; previously (as I understand it) the
Subject CN entry was used instead, but this is now ignored.

There doesn't appear to be any way to specify this directly in the `openssl` command, so
I pulled out all the config into a config file.

See https://textslashplain.com/2017/03/10/chrome-deprecates-subject-cn-matching/.